requirements.txt
pillow
stepic
pycryptodome
tk
